@@ -1,48 +1,56 @@
 {
-  "name": "@lopatnov/worker-from-string",
-  "version": "1.3.2",
+  "name": "@lopatnov/sample-library",
+  "version": "1.0.0",
   "license": "Apache-2.0",
   "author": "lopatnov",
-  "description": "Creates worker from string variable in a browser",
-  "homepage": "https://lopatnov.github.io/worker-from-string/",
+  "description": "Demonstrates a sample library template",
+  "homepage": "https://my-templates.github.io/TypeScript-Library/",
   "keywords": [
     "TypeScript",
     "JavaScript",
-    "Worker"
+    "library",
+    "typescript-library",
+    "javascript-library"
   ],
-  "main": "dist/worker-from-string.js",
-  "module": "dist/worker-from-string.es.js",
-  "typings": "dist/worker-from-string.d.ts",
+  "umdName": "sampleLibrary",
+  "libraryFile": "library",
+  "main": "dist/library.js",
+  "module": "dist/library.es.js",
+  "types": "dist/library.d.ts",
   "files": [
     "dist"
   ],
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/lopatnov/worker-from-string.git"
+    "url": "git+https://github.com/my-templates.github.io/TypeScript-Library.git"
   },
   "bugs": {
-    "url": "https://github.com/lopatnov/worker-from-string/issues"
+    "url": "https://github.com/my-templates.github.io/TypeScript-Library/issues"
   },
   "scripts": {
-    "build": "tsc --module commonjs && rollup -c rollup.config.ts",
-    "test": "npm run build && node run-tests.js",
-    "prepare": "npm i --no-save --no-package-lock terser",
+    "build": "rollup -c rollup.config.ts",
+    "watch": "rollup -cw rollup.config.ts",
     "configure-npm": "node ./set-registry.js -s https://registry.npmjs.org",
-    "configure-gpr": "node ./set-registry.js -s https://npm.pkg.github.com"
+    "configure-gpr": "node ./set-registry.js -s https://npm.pkg.github.com",
+    "test": "jest"
   },
   "publishConfig": {
     "access": "public"
   },
   "devDependencies": {
-    "@lopatnov/rollup-plugin-uglify": "^1.2.0",
-    "@types/node": "^12.11.7",
-    "express": "^4.17.1",
-    "rollup": "^1.25.2",
+    "@lopatnov/rollup-plugin-uglify": "^1.2.1",
+    "@types/jest": "^24.0.25",
+    "@types/typescript": "^2.0.0",
+    "jest": "^24.9.0",
+    "jest-config": "^24.9.0",
+    "rollup": "^1.27.14",
     "rollup-plugin-commonjs": "^10.1.0",
     "rollup-plugin-json": "^4.0.0",
     "rollup-plugin-node-resolve": "^5.2.0",
     "rollup-plugin-sourcemaps": "^0.4.2",
-    "rollup-plugin-typescript2": "^0.24.3",
-    "typescript": "^3.6.4"
+    "rollup-plugin-typescript2": "^0.25.3",
+    "terser": "^4.4.3",
+    "ts-jest": "^24.2.0",
+    "typescript": "^3.7.4"
   }
 }
