@@ -1,26 +1,19 @@
-import resolve from "rollup-plugin-node-resolve";
-import commonjs from "rollup-plugin-commonjs";
-import sourceMaps from "rollup-plugin-sourcemaps";
-import typescript from "rollup-plugin-typescript2";
 import json from "rollup-plugin-json";
+import typescript from "rollup-plugin-typescript2";
+import commonjs from "rollup-plugin-commonjs";
+import resolve from "rollup-plugin-node-resolve";
 import uglify from "@lopatnov/rollup-plugin-uglify";
 
-const pkg = require("./package.json");
+import pkg from "./package.json";
 
 export default [
   {
-    input: "src/worker-from-string.ts",
+    input: `src/${pkg.libraryFile}.ts`,
     output: [
       {
         file: pkg.main,
-        name: "workerFromString",
         format: "umd",
-        sourcemap: true
-      },
-      {
-        file: "test/worker-from-string.js",
-        name: "workerFromString",
-        format: "umd",
+        name: pkg.umdName,
         sourcemap: true
       },
       {
@@ -29,45 +22,38 @@ export default [
         sourcemap: true
       }
     ],
-    external: [...Object.keys(pkg.peerDependencies || {})],
-    watch: {
-      include: "src/**/*"
-    },
-    plugins: [
-      json(),
-      typescript({
-        typescript: require("typescript"),
-        useTsconfigDeclarationDir: true
-      }),
-      commonjs(),
-      resolve(),
-      sourceMaps()
-    ]
-  },
-  {
-    input: "src/worker-from-string.ts",
-    output: [
-      {
-        file: "test/worker-from-string.min.js",
-        name: "workerFromString",
-        format: "umd"
-      }
-    ],
     external: [
       ...Object.keys(pkg.devDependencies || {}),
       ...Object.keys(pkg.peerDependencies || {})
     ],
-    watch: {
-      include: "src/**/*"
-    },
     plugins: [
       json(),
       typescript({
-        typescript: require("typescript"),
-        useTsconfigDeclarationDir: true
+        typescript: require("typescript")
       }),
-      commonjs(),
       resolve(),
+      commonjs()
+    ]
+  },
+  {
+    input: `src/${pkg.libraryFile}.ts`,
+    output: {
+      file: `dist/${pkg.libraryFile}.min.js`,
+      name: pkg.umdName,
+      format: "umd",
+      sourcemap: true
+    },
+    external: [
+      ...Object.keys(pkg.devDependencies || {}),
+      ...Object.keys(pkg.peerDependencies || {})
+    ],
+    plugins: [
+      json(),
+      typescript({
+        typescript: require("typescript")
+      }),
+      resolve(),
+      commonjs(),
       uglify()
     ]
   }
